Move models to /models
